Fixing how the mood is modify, fixing GameManager prefab which had broken component it it 🤔

+++ Assets/Scripts/System/Temporality.cs
@@ -30,7 +30,7 @@ public class Temporality : MonoBehaviour {
 
     public float GetMicroCoef()
     {
-        return 1 / nbMicroCyclePerCycle;
+        return 1f / (float)nbMicroCyclePerCycle;
     }
 
     public void PauseTime()
